requirements.txt
PyMuPDF
openai
azure-identity
azure-search-documents==11.4.0b6
azure-storage-blob
tenacity
pandas
tiktoken
langchain
pyodbc
sqlalchemy
streamlit
